HYD-5566 Add slashes to source and test directories in npm ignore to prevent recursive matching
Obj is not currently publishing with the "dist" directory like it
should. To resolve this, the "source" and "test" folders need to have a
"/" in front of them in the .npmignore file. This will prevent npm from
reading the directories recursively and will thus pay attention to them
at the root level.

This patch will update the .npmigore file to include the slashes for
both the source and test folders.

- Added slashes to appropriate directories in .npmignore